+++ src/stringAlgorithms.cpp
@@ -8,16 +8,11 @@
 
 #include "stringAlgorithms.h"
 
-//----------------------------
-//Returns a vector containing the zero based index of
-//  the start of each match of the string K in S.
-//  Matches may overlap
-//----------------------------
 vector<int> StringAlgorithms::KMP(string S, string K)
 {
     vector<int> T(K.size() + 1, -1);
     vector<int> matches;
-    
+
     if(K.size() == 0)
     {
         matches.push_back(0);
@@ -29,7 +24,7 @@ vector<int> StringAlgorithms::KMP(string S, string K)
         while(pos != -1 && K[pos] != K[i - 1]) pos = T[pos];
         T[i] = pos + 1;
     }
-    
+
     int sp = 0;
     int kp = 0;
     while(sp < S.size())
@@ -37,12 +32,11 @@ vector<int> StringAlgorithms::KMP(string S, string K)
         while(kp != -1 && (kp == K.size() || K[kp] != S[sp])) kp = T[kp];
         kp++;
         sp++;
-        if(kp == K.size()) matches.push_back(sp - K.size());
+        if(kp == K.size()) matches.push_back(sp - (int)K.size());
     }
-    
+
     return matches;
 }
-
 // Boyer moore came from 
 // http://www-igm.univ-mlv.fr/~lecroq/string/node14.html#SECTION0014
